+++ src/transfer.py
@@ -1,5 +1,4 @@
 from PIL import Image, ImageChops, ImageColor, ImageEnhance
-import argparse
 import collections
 import os
 import glob
@@ -15,12 +14,15 @@ class Transfer:
     colour = None
     merge = 0
     brightness = 1.0
+    contrast = 1.0
+    sharpness = 1.0
     destination = None
     base = ""
     goal = ""
     base_filter = "*"
     goal_filter = "*"
     out_path = ""
+    size = None
     verbose = True
     exclude = []
     include = []
@@ -32,7 +34,7 @@ class Transfer:
         colors = self.__get_colors_by_frequency(value)
         return colors.most_common(1)[0][0]
 
-    def __tint_image(self, dest:object, goal:object, colour=None) -> object:
+    def __tint_image(self, dest: object, goal: object, colour=None) -> object:
         tint_color = ()
         if colour is None:
             pixel_values = goal.load()
@@ -48,11 +50,13 @@ class Transfer:
         pixel_values = image.load()
         return image, width, height, pixel_values
 
-    def __outName(self, image_path: Path, out_path: str) -> str:
+    def __outName(self, image_path: Path, out_path: str, width: int = None, height: int = None) -> str:
         name = os.path.basename(image_path).split('.')[0]
+        if height and width:
+            name += f"_{width}x{height}"
         return os.path.join(out_path, name)
 
-    def __create_dirs(self, out_path:str) -> None:
+    def __create_dirs(self, out_path: str) -> None:
         try:
             os.makedirs(out_path)
         except OSError as exc:
@@ -61,7 +65,7 @@ class Transfer:
             else:
                 raise
 
-    def __alpha_transfer(self, base_values:object, goal_values:object) -> Image:
+    def __alpha_transfer(self, base_values: object, goal_values: object) -> Image:
         destination_image = Image.new('RGBA', (self.width, self.height))
         destination_values = destination_image.load()
 
@@ -79,15 +83,20 @@ class Transfer:
                 destination_values[i, j] = tuple(base)
 
         return destination_image
-    
+
+    def __resizeImage(self, image: Image, size: tuple) -> Image:
+        assert len(
+            size) == 2, f"size needs be a tuple (width, height) not {size}"
+        return image.resize(size, Image.ANTIALIAS)
+
     def __saveImage(self, goal_image_path: Path, destination_image: Image, out_path: str) -> None:
-        out = self.__outName(goal_image_path, out_path)
+        out = self.__outName(goal_image_path, out_path,
+                             self.width, self.height)
         out = f"{out}.png"
         if self.verbose:
             print(f"Save to {out}")
         destination_image.save(out)
 
-    
     def createImage(self, base_image_path: Path, goal_image_path: Path, out_path="out") -> None:
         self.__create_dirs(out_path)
 
@@ -96,16 +105,16 @@ class Transfer:
         goal_image, goal_width, goal_height, goal_values = self.__imageInformations(
             goal_image_path)
 
-        assert base_height == goal_height, "Base and goal image need to have the same height"
-        assert base_width == goal_width, "Base and goal image need to have the same height"
-        #assert len(goal_values[0,0]) == 4, "Goal image needs to have an alpha pixel"
-
         if base_image.mode not in ['RGB', 'RGBA']:
             raise TypeError(f'Unsupported base image mode: {base_image.mode}')
         if goal_image.mode not in ['RGBA']:
             raise TypeError(f'Unsupported goal image mode: {goal_image.mode}')
 
         self.height, self.width = base_height, base_width
+
+        if goal_height != base_height or goal_width != base_width:
+            goal_image = self.__resizeImage(
+                goal_image, (self.width, self.height))
 
         destination_image = self.__alpha_transfer(base_values, goal_values)
 
@@ -116,9 +125,20 @@ class Transfer:
         destination_image = ImageEnhance.Brightness(
             destination_image).enhance(self.brightness)
 
+        destination_image = ImageEnhance.Contrast(
+            destination_image).enhance(self.contrast)
+
+        destination_image = ImageEnhance.Sharpness(
+            destination_image).enhance(self.sharpness)
+
+        if self.size is not None:
+            self.width, self.height = self.size
+            destination_image = self.__resizeImage(
+                destination_image, self.size)
+
         self.__saveImage(goal_image_path, destination_image, out_path)
 
-    def __SkipImage(self, image:Path):
+    def __SkipImage(self, image: Path):
         for e in self.exclude:
             print(image, e)
             if e in str(image):
@@ -130,7 +150,7 @@ class Transfer:
                 break
         return isNotIncluded
 
-    def __setupGoal(self, base_image:str, out_path:str):
+    def __setupGoal(self, base_image: str, out_path: str):
         if os.path.isdir(self.goal):
             for goal_image in Path(self.goal).rglob(self.goal_filter):
                 if self.__SkipImage(goal_image):
@@ -154,12 +174,14 @@ class Transfer:
             out_path = self.__outName(self.base, self.out_path)
             self.__setupGoal(self.base, out_path)
 
-    def __init__(self, base_path: str, goal_path: str, 
-        base_filter="*", goal_filter="*", exclude=[], include=[], out_path="out", 
-        verbose=True, tint=False, colour=None, merge=0, brightness=1.0, tint_threshold=100):
+    def __init__(self, base_path: str, goal_path: str,
+                 base_filter="*", goal_filter="*", exclude=[], include=[], out_path="out", size=None,
+                 verbose=True, tint=False, colour=None, merge=0, brightness=1.0, contrast=1.0, sharpness=1.0, tint_threshold=100):
         assert tint_threshold <= 255, "Threshold is too big"
         assert brightness >= 0.0, "Brightness is to small"
         assert merge >= 0, "Merge must be positive"
+        assert size is None or len(
+            size) == 2, "Size needs to be a tuple of 2 values: (width, height)"
         self.base = base_path
         self.goal = goal_path
         self.goal_filter = goal_filter
@@ -168,14 +190,19 @@ class Transfer:
         self.exclude = exclude
         self.include = include
         self.verbose = verbose
+        self.size = size
         self.tint = tint
         self.colour = colour
         self.merge = merge
         self.brightness = brightness
+        self.contrast = contrast
+        self.sharpness = sharpness
         self. tint_threshold = tint_threshold
 
 
 def get_parser():
+    import argparse
+
     def check_100(value: int):
         ivalue = int(value)
         if ivalue < 0 or ivalue > 100:
@@ -188,45 +215,73 @@ def get_parser():
     parser.add_argument('base', help="Image you want to see edited.")
     parser.add_argument(
         'goal', help="Image you want to take the alpha value and tint from.")
-    parser.add_argument('-bf', '--base_filter', default="*", help="Filter the base image folder.")
-    parser.add_argument('-gf', '--goal_filter', default="*", help="Filter the goal image folder.")
-    a2 = parser.add_mutually_exclusive_group(required=False)
-    a2.add_argument('-e', '--exclude', nargs='*', default=[], help="Exclude a list of file names.")
-    a2.add_argument('-i', '--include', nargs='*', default=[], help="Include only a list file names.")
-    a2.add_argument('-o', '--out', default="out", help="Base directory where the output is saved.")
+    ag0 = parser.add_argument_group(
+        'Filter the source images in a folder.')
+    ag0.add_argument('-bf', '--base_filter', default="*",
+                        help="Filter the base image folder.")
+    ag0.add_argument('-gf', '--goal_filter', default="*",
+                        help="Filter the goal image folder.")
+    e1 = ag0.add_mutually_exclusive_group(required=False)
+    e1.add_argument('-e', '--exclude', nargs='*', default=[],
+                    help="Exclude a list of file names.")
+    e1.add_argument('-i', '--include', nargs='*', default=[],
+                    help="Include only a list file names.")
+    parser.add_argument('-o', '--out', default="out",
+                        help="Base directory where the output is saved.")
+    parser.add_argument('-s', '--size', nargs='+', default=None,
+                        type=int, help="Width and height of the resulting image.")
     parser.add_argument('-m', '--merge', type=check_100,
                         help="Merge with goal in percent from 0-100.", default=0)
-    parser.add_argument('-b', '--brightness',
-                        help="Improve brightness with an value between 0.0 and infty. 0.0 is black, 1.0 is normal and everything above is brighter.", type=float, default=1.0)
     parser.add_argument('-v', '--verbose', action='store_true', default=False)
-    ag1 = parser.add_argument_group('Tint the base images with the majority colour of the goal image or a given colour.')
-    ag1.add_argument('-t', '--tint', help="Add tint from the goal image.",
+    ag1 = parser.add_argument_group(
+        'Enhance various aspects of the image.')
+    ag1.add_argument('-b', '--brightness',
+                     help="Improve brightness with an value between 0.0 and infty. 0.0 is black, 1.0 is normal and everything above is brighter.", type=float, default=1.0)
+    ag1.add_argument('-ct', '--contrast',
+                     help="Improve contrast with an value between 0.0 and infty.", type=float, default=1.0)
+    ag1.add_argument('-sh', '--sharpness',
+                     help="Improve sharpness with an value between 0.0 and infty.", type=float, default=1.0)
+    ag2 = parser.add_argument_group(
+        'Tint the base images with the majority colour of the goal image or a given colour.')
+    ag2.add_argument('-t', '--tint', help="Add tint from the goal image.",
                      action='store_true', default=False)
-    ag1.add_argument('-c', '--colour', help="Specify a colour.")
+    ag2.add_argument('-c', '--colour', help="Specify a colour.")
 
     return parser
 
+
+
+def __add_to_name(name, value, needSeparator, out_path):
+    if needSeparator:
+        out_path += "_"
+    out_path += f"{name}{value}"
+    return True, out_path
+
 if __name__ == "__main__":
 
-    args = get_parser().parse_args()
+    parser = get_parser()
+    args = parser.parse_args()
 
     out_path = os.path.join(args.out, "")
+    needSeparator = False
+
     if args.tint:
         out_path += "t"
         if args.colour is not None:
             out_path += f"_{args.colour}"
+        needSeparator = True
     if args.brightness != 1.0:
-        if args.tint:
-            out_path += "_"
-        out_path += f"b{args.brightness}"
+        needSeparator, out_path = __add_to_name("b", args.brightness, needSeparator, out_path)
+    if args.contrast != 1.0:
+        needSeparator, out_path = __add_to_name("c", args.contrast, needSeparator, out_path)
+    if args.sharpness != 1.0:
+        needSeparator, out_path = __add_to_name("s", args.sharpness, needSeparator, out_path)
     if args.merge > 0:
-        if args.tint or args.brightness != 1.0:
-            out_path += "_"
-        out_path += f"m{args.merge}"
+        needSeparator, out_path = __add_to_name("m", args.merge, needSeparator, out_path)
 
     if args.verbose:
         print(out_path)
 
     transfer = Transfer(args.base, args.goal, args.base_filter, args.goal_filter, args.exclude,
-                        args.include, out_path, args.verbose, args.tint, args.colour, args.merge, args.brightness)
+                        args.include, out_path, args.size, args.verbose, args.tint, args.colour, args.merge, args.brightness, args.contrast, args.sharpness)
     transfer.setupBase()